+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.3.0 - New Commands!
+- "Add Object Reference to Deploy File" & "Refresh Object References in Deploy File" commands are now available
+- "Add Object Reference to Deploy File" allows users to select an XML file from their project to automatically add to the `deploy.xml` file. Once importing an object through the SuiteCloud Development Framework extension, use this command to quickly add the object reference!
+- "Refresh Object References in Deploy File" allows users to ensure that the file path of the objects defined in the `deploy.xml` file are correct. Once organising your files, you can run this command to quickly update the file path if it has changed since adding to the file! This will also remove any object references that do not exist in your current project, making it easy to ensure that you are only referencing objects that actually exist.
+
 ## 1.2.2
 - Bug fixes